Backend/requirements.txt
Flask==1.0.2
Flask-sqlalchemy==2.3.2
gunicorn==19.8.1
Keras==2.2.0
psycopg2==2.7.5
pycrypto==2.6.1
selenium==3.13.0
sqlalchemy==1.2.9
